ai/pyproject.toml
[tool.poetry]
name = "ml"
version = "0.1.0"
description = ""
authors = ["Sukhrob Ilyosbekov <SuxrobGM@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.112.1"}
pydicom = "^2.4.4"
pillow = "^10.4.0"
scikit-image = "^0.24.0"
streamlit = "^1.37.1"
matplotlib = "^3.9.2"
torch = {version = "^2.4.0+cu124", source = "pytorch_cuda"}
torchvision = {version = "^0.19.0+cu124", source = "pytorch_cuda"}
torchaudio = {version = "^2.4.0+cu124", source = "pytorch_cuda"}
torchxrayvision = "^1.2.3"
lightning = "^2.4.0"
tensorflow = "^2.17.0"
python-dotenv = "^1.0.1"
python-multipart = "^0.0.9"
opencv-python-headless = "^4.10.0.84"


[[tool.poetry.source]]
name = "pytorch_cuda"
url = "https://download.pytorch.org/whl/cu124"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
